refactor(storefront/files): move routes file
